third_party/WebKit/LayoutTests/http/tests/w3c/webperf/approved/navigation-timing/html/test_navigation_type_backforward-expected.txt

frame "<!--framePath //<!--frame0-->-->" - has 1 onunload handler(s)
frame "<!--framePath //<!--frame0-->-->" - has 1 onunload handler(s)
frame "<!--framePath //<!--frame0-->-->" - has 1 onunload handler(s)
frame "<!--framePath //<!--frame0-->-->" - has 1 onunload handler(s)

 This is a testharness.js-based test.
PASS window.performance is defined 
PASS window.performance.navigation is defined 
PASS window.performance.navigation.type == TYPE_NAVIGATE 
PASS window.performance.navigation.type == TYPE_BACK_FORWARD after history.back() 
PASS window.performance.navigation.type == TYPE_BACK_FORWARD after history.forward() 
Harness: the test ran to completion.
